backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
langchain==0.1.0
langchain-anthropic==0.1.0
anthropic>=0.16.0
pydantic==2.5.0
python-multipart==0.0.6